+++ tests/Core/Framework/Webhook/_fixtures/BusinessEvents/EntityBusinessEvent.php
@@ -27,8 +27,21 @@ class EntityBusinessEvent implements BusinessEventInterface, BusinessEventEncode
             ->add('tax', new EntityType(TaxDefinition::class));
     }
 
-    public function getEncodeValues(): array
+    public function getEncodeValues(string $shopwareVersion): array
     {
+        if (version_compare($shopwareVersion, '6.3.0.0', '<')) {
+            $createdAt = $this->tax->getCreatedAt()->format(DATE_ATOM);
+            $foreignKeys = [
+                'extensions' => [],
+            ];
+        } else {
+            $createdAt = $this->tax->getCreatedAt()->format(DATE_RFC3339_EXTENDED);
+            $foreignKeys = [
+                'extensions' => [],
+                'apiAlias' => null,
+            ];
+        }
+
         return [
             'tax' => [
                 'id' => $this->tax->getId(),
@@ -39,12 +52,10 @@ class EntityBusinessEvent implements BusinessEventInterface, BusinessEventEncode
                 'products' => null,
                 'customFields' => null,
                 'translated' => [],
-                'createdAt' => $this->tax->getCreatedAt()->format(DATE_ATOM),
+                'createdAt' => $createdAt,
                 'updatedAt' => null,
                 'extensions' => [
-                    'foreignKeys' => [
-                        'extensions' => [],
-                    ],
+                    'foreignKeys' => $foreignKeys,
                 ],
                 'apiAlias' => 'tax',
             ],