+++ tests/Core/Framework/Webhook/_fixtures/BusinessEvents/StructuredArrayObjectBusinessEvent.php
@@ -33,7 +33,7 @@ class StructuredArrayObjectBusinessEvent implements BusinessEventInterface, Busi
             );
     }
 
-    public function getEncodeValues(): array
+    public function getEncodeValues(string $shopwareVersion): array
     {
         return [
             'inner' => [